# go-server/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Tracker API
    version: 0.0.1
paths:
    /createblock:
        post:
            tags:
                - Tracker
            operationId: Tracker_CreateBlockFunc
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateBlock'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BlockResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /deleteblock/{UserId}/{BlockId}:
        delete:
            tags:
                - Tracker
            operationId: Tracker_DeletBlockFunc
            parameters:
                - name: UserId
                  in: path
                  required: true
                  schema:
                    type: string
                - name: BlockId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /getblocks/{UserId}:
        get:
            tags:
                - Tracker
            operationId: Tracker_ListBlockFunc
            parameters:
                - name: UserId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BlockResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /updateblock:
        put:
            tags:
                - Tracker
            operationId: Tracker_UpdateBlockFunc
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateBlock'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        BlockResp:
            type: object
            properties:
                BlockId:
                    type: string
                StartTime:
                    type: string
                StopTime:
                    type: string
                Duration:
                    type: string
                Label:
                    type: string
        CreateBlock:
            type: object
            properties:
                UserId:
                    type: string
                StartTime:
                    type: string
                Lapse:
                    type: string
                Duration:
                    type: string
                StopTime:
                    type: string
                Label:
                    type: string
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        UpdateBlock:
            type: object
            properties:
                UserId:
                    type: string
                BlockId:
                    type: string
        UpdateResponse:
            type: object
            properties:
                BlockId:
                    type: string
                Label:
                    type: string
                Duration:
                    type: string
                StartTime:
                    type: string
                StopTime:
                    type: string
tags:
    - name: Tracker